program3.java
In this program, the `DisplayFactors` method is used to find and print all the **factors of a given number**. The method accepts a number `iNo` as input and uses a `for` loop to check every number from `1` to `iNo`. During each iteration, the loop checks if `iNo` is divisible by the current value of `iCnt` using the modulus operator (`iNo % iCnt == 0`). If the remainder is zero, it means that `iCnt` divides the number completely and is therefore a factor. That factor is then printed on the screen. In the `main` method, an object of the `Logic` class calls the `DisplayFactors` method with input `12`, so the program prints all factors of 12 (i.e., 1, 2, 3, 4, 6, and 12).